+++ wiselib.testing/algorithms/tracking/PLTT_trace_revision.h
@@ -54,9 +54,9 @@ namespace wiselib
 			recipient_2_id		(0)
 		{
 			max_intensity = 2;
-			for (uint8_t i = 0; i < ( sizeof( max_intensity )*8 ); i++ )
+			for (size_t i = 0; i < ( sizeof( max_intensity ) * 8 ); i++ )
 			{
-				max_intensity = 2*max_intensity;
+				max_intensity = 2 * max_intensity;
 			}
 			max_intensity = max_intensity - 1;
 		}
@@ -69,16 +69,16 @@ namespace wiselib
 			start_time = _st;
 			inhibited = 0;
 			max_intensity = 2;
-			for (uint8_t i = 0; i < ( sizeof( max_intensity )*8 ); i++ )
+			for ( size_t i = 0; i < ( sizeof( max_intensity ) * 8 ); i++ )
 			{
-				max_intensity = 2*max_intensity;
+				max_intensity = 2 * max_intensity;
 			}
 			max_intensity = max_intensity - 1;
 			recipient_1_id = 0;
 			recipient_2_id = 0;
 			target_id = 0;
 		}
-		inline PLTT_TraceType( const self_type& _t)
+		inline PLTT_TraceType( const self_type& _t )
 		{
 			*this = _t;
 		}
@@ -87,7 +87,7 @@ namespace wiselib
 			get_from_buffer( buff, offset );
 			inhibited = 0;
 			max_intensity = 2;
-			for (uint8_t i = 0; i < ( sizeof( max_intensity )*8 ); i++ )
+			for (size_t i = 0; i < ( sizeof( max_intensity )*8 ); i++ )
 			{
 				max_intensity = 2*max_intensity;
 			}
@@ -95,17 +95,17 @@ namespace wiselib
 		}
 		inline block_data_t* set_buffer_from( block_data_t* buff, size_t offset = 0 )
 		{
-			uint8_t CURRENT_NODE_POS = 0;
-			uint8_t PARENT_NODE_POS = current.get_buffer_size() + CURRENT_NODE_POS;
-			uint8_t GRANDPARENT_NODE_POS = parent.get_buffer_size() + PARENT_NODE_POS;
-			uint8_t TARGET_ID_POS = grandparent.get_buffer_size() + GRANDPARENT_NODE_POS;
-			uint8_t START_TIME_POS = sizeof( NodeID ) + TARGET_ID_POS;
-			uint8_t DIMINISH_SECS_POS = sizeof( TimesNumber ) + START_TIME_POS;
-			uint8_t DIMINISH_AMOUNT_POS = sizeof( SecondsNumber ) + DIMINISH_SECS_POS;
-			uint8_t SPREAD_PENALTY_POS = sizeof( IntensityNumber ) + DIMINISH_AMOUNT_POS;
-			uint8_t INTENSITY_POS = sizeof( IntensityNumber ) + SPREAD_PENALTY_POS;
-			uint8_t RECIPIENT_1_ID_POS = sizeof( IntensityNumber ) + INTENSITY_POS;
-			uint8_t RECIPIENT_2_ID_POS = sizeof( NodeID ) + RECIPIENT_1_ID_POS;
+			size_t CURRENT_NODE_POS = 0;
+			size_t PARENT_NODE_POS = current.get_buffer_size() + CURRENT_NODE_POS;
+			size_t GRANDPARENT_NODE_POS = parent.get_buffer_size() + PARENT_NODE_POS;
+			size_t TARGET_ID_POS = grandparent.get_buffer_size() + GRANDPARENT_NODE_POS;
+			size_t START_TIME_POS = sizeof( NodeID ) + TARGET_ID_POS;
+			size_t DIMINISH_SECS_POS = sizeof( TimesNumber ) + START_TIME_POS;
+			size_t DIMINISH_AMOUNT_POS = sizeof( SecondsNumber ) + DIMINISH_SECS_POS;
+			size_t SPREAD_PENALTY_POS = sizeof( IntensityNumber ) + DIMINISH_AMOUNT_POS;
+			size_t INTENSITY_POS = sizeof( IntensityNumber ) + SPREAD_PENALTY_POS;
+			size_t RECIPIENT_1_ID_POS = sizeof( IntensityNumber ) + INTENSITY_POS;
+			size_t RECIPIENT_2_ID_POS = sizeof( NodeID ) + RECIPIENT_1_ID_POS;
 			current.set_buffer_from( buff, CURRENT_NODE_POS + offset );
 			parent.set_buffer_from( buff, PARENT_NODE_POS + offset);
 			grandparent.set_buffer_from( buff, GRANDPARENT_NODE_POS + offset);
@@ -122,17 +122,17 @@ namespace wiselib
 		// --------------------------------------------------------------------
 		inline void get_from_buffer(block_data_t* buff, size_t offset = 0)
 		{
-			uint8_t CURRENT_NODE_POS = 0;
-			uint8_t PARENT_NODE_POS = current.get_buffer_size() + CURRENT_NODE_POS;
-			uint8_t GRANDPARENT_NODE_POS = parent.get_buffer_size() + PARENT_NODE_POS;
-			uint8_t TARGET_ID_POS = grandparent.get_buffer_size() + GRANDPARENT_NODE_POS;
-			uint8_t START_TIME_POS = sizeof( NodeID ) + TARGET_ID_POS;
-			uint8_t DIMINISH_SECS_POS = sizeof( TimesNumber ) + START_TIME_POS;
-			uint8_t DIMINISH_AMOUNT_POS = sizeof( SecondsNumber ) + DIMINISH_SECS_POS;
-			uint8_t SPREAD_PENALTY_POS = sizeof( IntensityNumber ) + DIMINISH_AMOUNT_POS;
-			uint8_t INTENSITY_POS = sizeof( IntensityNumber ) + SPREAD_PENALTY_POS;
-			uint8_t RECIPIENT_1_ID_POS = sizeof( IntensityNumber ) + INTENSITY_POS;
-			uint8_t RECIPIENT_2_ID_POS = sizeof( NodeID ) + RECIPIENT_1_ID_POS;
+			size_t CURRENT_NODE_POS = 0;
+			size_t PARENT_NODE_POS = current.get_buffer_size() + CURRENT_NODE_POS;
+			size_t GRANDPARENT_NODE_POS = parent.get_buffer_size() + PARENT_NODE_POS;
+			size_t TARGET_ID_POS = grandparent.get_buffer_size() + GRANDPARENT_NODE_POS;
+			size_t START_TIME_POS = sizeof( NodeID ) + TARGET_ID_POS;
+			size_t DIMINISH_SECS_POS = sizeof( TimesNumber ) + START_TIME_POS;
+			size_t DIMINISH_AMOUNT_POS = sizeof( SecondsNumber ) + DIMINISH_SECS_POS;
+			size_t SPREAD_PENALTY_POS = sizeof( IntensityNumber ) + DIMINISH_AMOUNT_POS;
+			size_t INTENSITY_POS = sizeof( IntensityNumber ) + SPREAD_PENALTY_POS;
+			size_t RECIPIENT_1_ID_POS = sizeof( IntensityNumber ) + INTENSITY_POS;
+			size_t RECIPIENT_2_ID_POS = sizeof( NodeID ) + RECIPIENT_1_ID_POS;
 			current.get_from_buffer( buff, CURRENT_NODE_POS + offset);
 			parent.get_from_buffer( buff, PARENT_NODE_POS + offset );
 			grandparent.get_from_buffer( buff, GRANDPARENT_NODE_POS + offset );
@@ -147,17 +147,17 @@ namespace wiselib
 		}
 		inline size_t get_buffer_size()
 		{
-			uint8_t CURRENT_NODE_POS = 0;
-			uint8_t PARENT_NODE_POS = current.get_buffer_size() + CURRENT_NODE_POS;
-			uint8_t GRANDPARENT_NODE_POS = parent.get_buffer_size() + PARENT_NODE_POS;
-			uint8_t TARGET_ID_POS = grandparent.get_buffer_size() + GRANDPARENT_NODE_POS;
-			uint8_t START_TIME_POS = sizeof( NodeID ) + TARGET_ID_POS;
-			uint8_t DIMINISH_SECS_POS = sizeof( TimesNumber ) + START_TIME_POS;
-			uint8_t DIMINISH_AMOUNT_POS = sizeof( SecondsNumber ) + DIMINISH_SECS_POS;
-			uint8_t SPREAD_PENALTY_POS = sizeof( IntensityNumber ) + DIMINISH_AMOUNT_POS;
-			uint8_t INTENSITY_POS = sizeof( IntensityNumber ) + SPREAD_PENALTY_POS;
-			uint8_t RECIPIENT_1_ID_POS = sizeof( IntensityNumber ) + INTENSITY_POS;
-			uint8_t RECIPIENT_2_ID_POS = sizeof( NodeID ) + RECIPIENT_1_ID_POS;
+			size_t CURRENT_NODE_POS = 0;
+			size_t PARENT_NODE_POS = current.get_buffer_size() + CURRENT_NODE_POS;
+			size_t GRANDPARENT_NODE_POS = parent.get_buffer_size() + PARENT_NODE_POS;
+			size_t TARGET_ID_POS = grandparent.get_buffer_size() + GRANDPARENT_NODE_POS;
+			size_t START_TIME_POS = sizeof( NodeID ) + TARGET_ID_POS;
+			size_t DIMINISH_SECS_POS = sizeof( TimesNumber ) + START_TIME_POS;
+			size_t DIMINISH_AMOUNT_POS = sizeof( SecondsNumber ) + DIMINISH_SECS_POS;
+			size_t SPREAD_PENALTY_POS = sizeof( IntensityNumber ) + DIMINISH_AMOUNT_POS;
+			size_t INTENSITY_POS = sizeof( IntensityNumber ) + SPREAD_PENALTY_POS;
+			size_t RECIPIENT_1_ID_POS = sizeof( IntensityNumber ) + INTENSITY_POS;
+			size_t RECIPIENT_2_ID_POS = sizeof( NodeID ) + RECIPIENT_1_ID_POS;
 			return RECIPIENT_2_ID_POS + sizeof( NodeID );
 		}
 		inline self_type& operator=( const self_type& _t)
